refactored ui components into their own packages
