+++ Code/src/main/java/nl/utwente/group10/ui/components/anchors/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all of the Anchor components used by the UI
+ */
+package nl.utwente.group10.ui.components.anchors;
