+++ Code/src/main/java/nl/utwente/group10/ui/components/lines/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains all Line related components of the UI.
+ */
+package nl.utwente.group10.ui.components.lines;
