+++ Code/src/main/resources/ui/ConnectionAnchor.fxml
@@ -1,3 +1,3 @@
-<?import nl.utwente.group10.ui.components.ConnectionAnchor?>
-<fx:root type="nl.utwente.group10.ui.components.ConnectionAnchor" xmlns:fx="http://javafx.com/fxml/" radius="10">
+<?import nl.utwente.group10.ui.components.anchors.ConnectionAnchor?>
+<fx:root type="nl.utwente.group10.ui.components.anchors.ConnectionAnchor" xmlns:fx="http://javafx.com/fxml/" radius="10">
 </fx:root>
