+++ Code/src/main/resources/ui/ConnectionLine.fxml
@@ -1,3 +1,3 @@
-<?import nl.utwente.group10.ui.components.ConnectionLine?>
-<fx:root type="nl.utwente.group10.ui.components.ConnectionLine" xmlns:fx="http://javafx.com/fxml/" styleClass="connection">
+<?import nl.utwente.group10.ui.components.lines.ConnectionLine?>
+<fx:root type="nl.utwente.group10.ui.components.lines.ConnectionLine" xmlns:fx="http://javafx.com/fxml/" styleClass="connection">
 </fx:root>
